+++ benchmark/benchmark-server/src/main/java/com/accenture/services/ReceiveOnly.java
@@ -18,7 +18,6 @@
 
 package com.accenture.services;
 
-import org.platformlambda.core.annotations.CoroutineRunner;
 import org.platformlambda.core.annotations.PreLoad;
 import org.platformlambda.core.models.TypedLambdaFunction;
 
@@ -26,7 +25,6 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
-@CoroutineRunner
 @PreLoad(route="network.one.way", instances=200, isPrivate = false)
 public class ReceiveOnly implements TypedLambdaFunction<Map<String, Object>, Map<String, Object>> {