+++ system/platform-core/src/main/java/org/platformlambda/core/annotations/CoroutineRunner.java
@@ -1,8 +1,0 @@
-package org.platformlambda.core.annotations;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface CoroutineRunner { }