+++ system/platform-core/src/main/java/org/platformlambda/core/system/EventEmitter.java
@@ -903,12 +903,6 @@ public class EventEmitter {
      * You can retrieve result from the future's
      * onSuccess(EventEnvelope event)
      * onFailure(Throwable timeoutException)
-     * <p>
-     * IMPORTANT: This is an asynchronous RPC using Future.
-     * You should NOT use this API in your function that will be run as a coroutine.
-     * <p>
-     * If you annotate your LambdaFunction using "CoroutineRunner", please use "CoroutineBridge" to bridge your Java
-     * code to a kotlin suspend function that calls "awaitRequest" in "FastRPC.kt".
      *
      * @param event to the target
      * @param timeout in milliseconds
@@ -925,12 +919,6 @@ public class EventEmitter {
      * You can retrieve result from the future's
      * onSuccess(EventEnvelope event)
      * onFailure(Throwable timeoutException)
-     * <p>
-     * IMPORTANT: This is an asynchronous RPC using Future.
-     * You should NOT use this API in your function that will be run as a coroutine.
-     * <p>
-     * If you annotate your LambdaFunction using "CoroutineRunner", please use "CoroutineBridge" to bridge your Java
-     * code to a kotlin suspend function that calls "awaitRequest" in "FastRPC.kt".
      *
      * @param event to the target
      * @param timeout in milliseconds