+++ CHANGELOG.md
@@ -8,6 +8,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 3.0.11, 4/23/2024
+
+Backport KernelThreadRunner from Mercury Composable 3.1
+
+### Added
+
+KernelThreadRunner annotation
+
+### Removed
+
+CoroutineRunner annotation
+
+### Changed
+
+Set default execution strategy to "coroutine". To tell the system to run a function
+using kernel thread pool, add the `KernelThreadRunner` annotation.
+
+---
 ## Version 3.0.10, 3/19/2024
 
 ### Added
@@ -21,7 +39,7 @@ N/A
 
 ### Changed
 
-1. Open sources library update (Spring Boot 3.2.4, Vertx 4.5.5, Netty 4.1.105.Final)
+1. Open sources library update (Spring Boot 3.2.4, Vertx 4.5.5)
 2. Improve AppConfigReader and ConfigReader to use the app-config-reader.yml file.
 
 ---